+++ scglobals/static/tx/src/app/collections/shipments.js
@@ -12,4 +12,17 @@ export default Backbone.Collection.extend({
   save: function(){
     return this.db.bulkDocs(this.toJSON());
   },
+  fetch: function(){
+    return this.db.query('shipments-by-date');
+  }
+  // Promise.all([
+  //   this.db.query('scbydate',{reduce: true}),
+  //   this.db.query('scbyvalue',{reduce: true})
+  //   ]).then((result)=>{
+  //     $("#transactions-summary").html(
+  //       tsTemplate({total_transactions: result[0][0], total_value: result[1][0]})
+  //     );
+  //   },(err)=>{
+  //     console.log(err);
+  //   });
 });