+++ scglobals/static/tx/src/app/services/initialqueries.js
@@ -9,10 +9,11 @@ export default class InitialQueries {
 
   saveDefaultQueries(){
     let querytypes = [
-      {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: 'doc.date'},
+      {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: 'doc.date,doc.total_value'},
+      // {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: '[doc.date,doc.from_location_name,doc.to_location_name],doc.total_value'},
       // for sum of total value (when working, use sum & skip "count" of shipments)
       {queryName: 'shipments-by-value', doc_type: 'shipment', queryEmit: 'doc._id, doc.total_value', reduce: '_sum'},
-      {queryName: 'shipments-by-count', doc_type: 'shipment', queryEmit: 'doc._id, doc.total_items', reduce: '_sum'}];
+      {queryName: 'shipments-by-count', doc_type: 'shipment', queryEmit: 'doc._id, doc.total_transactions', reduce: '_sum'}];
     var addQueries = [];
     var addQuery;
     var functionString;