+++ scglobals/static/tx/src/app/views/dash.js
@@ -3,31 +3,27 @@ import _ from 'lodash';
 import Backbone from 'backbone';
 
 import DashTemplate from './../templates/dash.hbs';
-import tsTemplate from './../templates/transactions-summary.hbs';
-import ShipmentsView from './shipments-view';
-import DB from './../services/db';
+import ShipmentsSummaryTemplate from './../templates/shipments-summary.hbs';
+import ShipmentsTemplate from './../templates/shipments.hbs';
+import ShipmentsCollection from './../collections/shipments';
 
 export default Backbone.View.extend({
 
   template: DashTemplate,
+  shipmentsTemplate: ShipmentsTemplate,
   el: '#container',
   initialize: function(){
-    this.db = new DB();
+    this.shipmentsCollection = new ShipmentsCollection();
+  },
+  showShipments: function(){
+    $("#loading").fadeOut('fast');
+    $("#shipments").removeClass('hidden').hide().fadeIn(1000);
   },
   render: function() {
     this.$el.html(this.template());
-    this.db.query('scbydate').then((result)=>{
-      $('#transactionsdiv').html(new ShipmentsView().render(result));
+    this.shipmentsCollection.fetch().then((shipments)=>{
+      $("#shipments").html(this.shipmentsTemplate(shipments));
+      this.showShipments();
     });
-    Promise.all([
-      this.db.query('scbydate',{reduce: true}),
-      this.db.query('scbyvalue',{reduce: true})
-      ]).then((result)=>{
-        $("#transactions-summary").html(
-          tsTemplate({total_transactions: result[0][0], total_value: result[1][0]})
-        );
-      },(err)=>{
-        console.log(err);
-      });
   },
 });